+++ pom.xml
@@ -178,9 +178,9 @@
 			<version>4.4.11</version>
 		</dependency>
 		<dependency>
-			<groupId>servlets.com</groupId>
-			<artifactId>cos</artifactId>
-			<version>05Nov2002</version>
+		    <groupId>servlets.com</groupId>
+		    <artifactId>cos</artifactId>
+		    <version>05Nov2002</version>
 		</dependency>
 		<!-- commons FileUpload -->
         <dependency>
