+++ src/main/java/com/tje/webapp/HomeController.java
@@ -1,8 +1,7 @@
 ﻿package com.tje.webapp;
 
-import java.io.File;
 import java.io.IOException;
-import java.util.*;
+import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -14,18 +13,27 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 
-import com.fasterxml.jackson.core.JsonParser;
-import com.oreilly.servlet.MultipartRequest;
-import com.oreilly.servlet.multipart.DefaultFileRenamePolicy;
-import com.tje.model.*;
-import com.tje.page.*;
-import com.tje.service.*;
-import com.tje.webapp.setting.PageingInfo;
+import com.tje.model.Board_Free;
+import com.tje.model.DetailBoardFreeView;
+import com.tje.model.SimpleBoardFreeView;
+import com.tje.service.AllItemListService;
+import com.tje.service.Board_freeService;
+import com.tje.service.Board_freeViewService;
+import com.tje.service.CommentAddService;
+import com.tje.service.CommentDeleteService;
+import com.tje.service.CommentSelectService;
+import com.tje.service.DetailBoardFreeViewService;
+import com.tje.service.DetailBoardFreeView_DeleteService;
+import com.tje.service.DetailBoardFreeView_UpdateService;
+import com.tje.service.ItemAddService;
+import com.tje.service.ItemViewCntUpdateService;
+import com.tje.service.ItemViewService;
+import com.tje.service.LikeAndDislikeService;
+import com.tje.service.SimpleBoardFreeViewSelectByDateDescService;
+import com.tje.service.SimpleBoardItemListCountCriteriaService;
+import com.tje.service.SimpleBoardItemListCriteriaService;
 
 
 @Controller
@@ -170,206 +178,7 @@ public class HomeController {
 	
 	
 	
-	@RequestMapping(value = {"/item","/item/{curPageNo}"})
-	public String Item(Model model,Criteria criteria,
-			@PathVariable(value="curPageNo", required = false) Integer curPageNo) {
-		
-		System.out.println(criteria.toString());
-		
-		if(curPageNo==null)
-			curPageNo=1;
-		
-		criteria.setPage(curPageNo);
-		criteria.setPerPageNum(6);
-		model.addAttribute("item_list", sbilcService.service(criteria));
-		PageMaker pageMaker=new PageMaker();
-		
-		pageMaker.setCri(criteria);
-		pageMaker.setTotalCount((int)sbilccService.service());
-		System.out.println(pageMaker.toString());
-		
-		model.addAttribute("curPageNo", curPageNo);
-		model.addAttribute("pageMaker", pageMaker);  // 게시판 하단의 페이징 관련, 이전페이지, 페이지 링크 , 다음 페이지
-		
-//		List<SimpleBoardItemView> result =  (List<SimpleBoardItemView>) ilService.service();
-//		model.addAttribute("item_list", result);
-		return "item";
-	}
 	
-	@GetMapping("/add_item")
-	public String add_item() {
-		return "add_item";
-	}
-	
-	@PostMapping(value = "/add_item", produces = "application/text; charset=utf8")
-	@ResponseBody
-	public String add_item(HttpServletRequest request) {
-		String dirPath=request.getSession().getServletContext().getRealPath("/resources/images");
-		System.out.println(dirPath);
-		File dir=new File(dirPath);
-		
-		if(!dir.exists())
-			dir.mkdirs();
-		
-		int size=10*1024*1024;
-		
-		MultipartRequest multipartRequest=null;
-		int result=0;
-		
-		try {
-			multipartRequest=new MultipartRequest(request, dirPath, size, "utf-8", new DefaultFileRenamePolicy());
-			
-			String member_id=multipartRequest.getParameter("member_id");
-			String title=multipartRequest.getParameter("title");
-			String content=multipartRequest.getParameter("content");
-			String strCategory=multipartRequest.getParameter("category");
-			int category=Integer.parseInt(strCategory);
-			int number=Integer.parseInt(multipartRequest.getParameter("number"));
-			String price=multipartRequest.getParameter("price");
-			String imageName=multipartRequest.getFilesystemName("image");
-			String orginName=multipartRequest.getOriginalFileName("image");
-			
-			Board_Item item=new Board_Item(0, 3, category, title, content, member_id, number, price, imageName, 0, null);
-			
-			result=(int) aiService.service(item);
-			System.out.println(result);
-			
-			if(result==1)
-				return "상품 추가가 완료되었습니다.";
-			
-		} catch (Exception e) {
-		}
-		
-		return "상품 추가 과정에서 문제가 발생하였습니다.";
-	}
-	
-	@GetMapping("/item_view/{board_id}")
-	public String item_view(Model model, 
-			@PathVariable(value = "board_id") Integer board_id,
-			HttpSession session) {
-		
-		DetailBoardItemView item=new DetailBoardItemView();
-		item.setBoard_id(board_id);
-		
-		Comment comment=new Comment();
-		comment.setBoard_id(board_id);
-		comment.setTopic(3);
-		
-		
-		Member login_member=(Member) session.getAttribute("login_member");
-		if(login_member==null)
-			model.addAttribute("btn_status", 0);
-		else {
-			String member_id=login_member.getMember_id();
-			
-			LikeAndDislike lad=new LikeAndDislike();
-			lad.setMember_id(member_id);
-			lad.setBoard_id(board_id);
-			lad.setTopic(3);
-			
-			LikeAndDislike result=(LikeAndDislike) ladService.selectOne(lad);
-			if(result==null)
-				model.addAttribute("btn_status", 0);
-			else
-				model.addAttribute("btn_status", result.getIs_like());
-		}
-		
-		
-		if((int)ivcuService.service(item)!=1)
-			return "redirect:error/item_view";
-		
-		model.addAttribute("searchedItem", ivService.service(item));
-		model.addAttribute("commentList", csSercie.service(comment));
-		
-		return "item_view";
-	}
-	
-	@PostMapping(value = "/comment_add")
-	@ResponseBody
-	public Comment comment_add(
-			Comment comment,
-			Model model,
-			HttpServletResponse response) {
-		
-		int r=(int) caService.service(comment);
-		comment.setComment_id(r);
-		
-		response.setContentType("application/json; charset=utf-8");
-		
-		if(r!=0) {
-			return comment;
-		}
-		
-		return null;
-	}
-	
-	@PostMapping(value = "/comment_delete", produces = "application/text; charset=utf8")
-	@ResponseBody
-	public String comment_delete(@RequestParam("comment_id") int comment_id) {
-		
-		Comment comment=new Comment();
-		comment.setComment_id(comment_id);
-		
-		int r=(int) cdService.service(comment);
-		
-		if(r==1) {
-			return "댓글 삭제을 완료했습니다.";
-		}
-		
-		return "댓글 삭제를 실패했습니다.";
-	}
-	
-	@PostMapping(value = "/like_and_dislike", produces = "application/text; charset=utf8")
-	@ResponseBody
-	public String like_and_dislike(
-			@RequestBody Map<String, Object> map,
-			HttpSession session,
-			Model model) {
-		
-		Member login_member=(Member) session.getAttribute("login_member");
-		
-		LikeAndDislike lad=new LikeAndDislike();
-		lad.setMember_id(login_member.getMember_id());
-		lad.setBoard_id((int) map.get("board_id"));
-		lad.setTopic((int) map.get("topic"));
-		
-		int status=(int) map.get("status");
-		
-		int r=0;
-		
-		switch (status) {
-		case 1:	
-			lad.setIs_like(1);
-			r=(int) ladService.update(lad);
-			break;
-		case 2:	
-			lad.setIs_like(1);
-			r=(int) ladService.insert(lad);
-			break;
-		case 3:	
-			r=(int) ladService.delete(lad);
-			break;
-		case 4:	
-			lad.setIs_like(2);
-			r=(int) ladService.update(lad);
-			break;
-		case 5:	
-			lad.setIs_like(2);
-			r=(int) ladService.insert(lad);
-			break;
-		case 6:	
-			r=(int) ladService.delete(lad);
-			break;
-		default:
-			break;
-		}
-		
-		if(r!=0) {
-			return "success";
-		}
-		
-		return "fail";
-	}
 	
 	@RequestMapping("/cart")
 	public String Cart() {
