+++ src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -33,13 +33,8 @@
 	<beans:bean id="naverLoginBO" class="com.tje.api.NaverLoginBO" />
 	<!-- kakao 로그인 -->
 	<beans:bean name="auth" class="com.tje.api.Auth" />
-	<!-- file Upload -->
-    <beans:bean id="multipartResolver" class="org.springframework.web.multipart.commons.CommonsMultipartResolver">
-        <!-- max upload size in bytes -->
-        <beans:property name="maxUploadSize" value="20971520" /> <!-- 20MB -->
-        <!-- max size of file in memory (in bytes) -->
-        <beans:property name="maxInMemorySize" value="1048576" /> <!-- 1MB -->
-    </beans:bean>
+
+
 	
 	
 </beans:beans>
